Create Stepper_Crtl for Linear Actuator
When the left or right control button is pressed, the NEMA 17 1.2-amp stepper motor will rotate accordingly. When neither or both are pressed, the motor will not rotate at all. 
The code is presently written to make the motor rotate at its fastest possible speed. This is determined by setting the rate at which the high/low cycle of the driver takes place (currently set to 450 microseconds). The minimum rate appears to be somewhere in the range of 400 microseconds. There does not appear to be an upper limit.